Added main for ft_substr and moved big main files to utils

+++ srcs/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: pedrohe3 <pedrohe3@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/22 18:41:36 by pedrohe3          #+#    #+#             */
-/*   Updated: 2025/10/22 18:54:08 by pedrohe3         ###   ########.fr       */
+/*   Updated: 2025/10/23 17:36:24 by pedrohe3         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int count)
+void	*ft_memcpy(void *dest, const void *src, size_t count)
 {
 	unsigned int	i;
