+++ srcs/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: pedrohe3 <pedrohe3@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 17:55:01 by pedrohe3          #+#    #+#             */
-/*   Updated: 2025/10/22 13:37:56 by pedrohe3         ###   ########.fr       */
+/*   Updated: 2025/10/23 16:03:54 by pedrohe3         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	idx;
