+++ srcs/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: pedrohe3 <pedrohe3@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/20 17:54:58 by pedrohe3          #+#    #+#             */
-/*   Updated: 2025/10/22 14:56:43 by pedrohe3         ###   ########.fr       */
+/*   Updated: 2025/10/23 17:42:06 by pedrohe3         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(char *s1, char *s2, unsigned int n)
+int	ft_strncmp(char *s1, char *s2, size_t n)
 {
 	unsigned int	idx;
